+++ core/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 1.2.7
 
+- Changed the default AutoNat `only_global_ips` from `false` to `true` so the servers (and clients) discard AutoNat probes
+  from the private-only peers
 - Updated the default AutoNat configuration with increased timeouts
 - Added counters for initial and switched RPC host connections to telemetry metrics
 - Added put records counter to telemetry metrics